requirements.txt
pytest>=7.3.0
selene>=2.0.0b16
allure-pytest>=2.13.2